requirements.txt
diffusers
transformers
safetensors
sentencepiece
optimum-quanto
protobuf
hf-transfer
accelerate
bitsandbytes
huggingface-hub
ftfy==6.1.3
scipy==1.11.4
numpy==1.26.2
opencv-python==4.7.0.68
tqdm==4.66.1
Pillow==10.1.0
timm==0.9.12
ipython==8.18.1
responses==0.24.1
httpx==0.25.2
hypothesis==6.92.0
python-dotenv
wcwidth
tabulate
pydantic==2.5.2
flask==3.0.0
waitress==2.1.2
nltk==3.8.1
pytorch-seed==0.2.0
invisible-watermark==0.2.0
pika==1.3.2
python-logging-loki==0.3.1
peft
aura-sr
runpod